Crash #1289263

+++ OpenSim/Addons/os-webrtc-janus/Janus/JanusViewerSession.cs
@@ -55,7 +55,7 @@ namespace WebRtcVoice
         public OMV.UUID AgentId { get; set; }
 
         // Janus keeps track of the user by this ID
-        public int ParticipantId { get; set; }
+        public long ParticipantId { get; set; }
 
         // Connections to the Janus server
         public JanusSession Session { get; set; }
